backend/requirements.txt
Django==4.0.2
djangorestframework==3.13.1
drf-spectacular==0.21.2
pytz
gunicorn==20.1.0
django-heroku==0.3.1
requests
cachetools==5.0.0
drf-extra-fields
Pillow==9.0.1
feedparser==6.0.8
bs4==0.0.1
sendgrid==6.9.6
python-dotenv